clean up and split chat components
