+++ packages/dashx-react/src/hooks/index.ts
@@ -1,4 +1,5 @@
 export { default as useAgent } from './use-agent.js';
+export { default as useAnimatedText } from './use-animated-text.js';
 export { default as useDashXProvider } from './use-dashx-provider.js';
 export { default as useInApp } from './use-in-app.js';
 export { default as useProductRelease } from './use-product-release.js';
