+++ packages/dashx-react/src/widgets/Chat.tsx
@@ -1,13 +1,12 @@
-import * as ScrollArea from '@radix-ui/react-scroll-area';
-import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { Send } from 'lucide-react';
-import type { AiAgent, AiMessage, AiAgentStarterSuggestion } from '@dashx/browser';
-import type { KeyboardEvent } from 'react';
+import React, { useState } from 'react';
 
 import { cn } from '../utils/cn.js';
-import { Button, Flex, Text, TextArea, Theme, Heading, Popover, MarkdownRenderer } from '../components';
+import { Button, Flex, Theme, Popover } from '../components';
 import { useAgent } from '../hooks';
 import { X } from '../icons';
+import ChatHeader from './ChatHeader.js';
+import ChatBody from './ChatBody.js';
+import ChatFooter from './ChatFooter.js';
 
 type ChatProps = {
   publicEmbedKey: string;
@@ -19,7 +18,6 @@ type ChatProps = {
 const Chat = ({ publicEmbedKey, withChatHeader = false, withPopoverClose = false, borderless = false }: ChatProps) => {
   const { agent, messages, isThinking, error, sendMessage } = useAgent({ publicEmbedKey });
   const [isAnimating, setIsAnimating] = useState(false);
-
 
   if (!agent) {
     return (
@@ -34,7 +32,6 @@ const Chat = ({ publicEmbedKey, withChatHeader = false, withPopoverClose = false
       </Theme>
     );
   }
-
 
   return (
     <Theme asChild>
@@ -63,190 +60,6 @@ const Chat = ({ publicEmbedKey, withChatHeader = false, withPopoverClose = false
     </Theme>
   );
 };
-
-type ChatHeaderProps = {
-  agent: AiAgent;
-  children?: React.ReactNode;
-};
-
-const ChatHeader = ({ agent, children }: ChatHeaderProps) => {
-  return (
-    <Flex
-      align="center"
-      className="border-b border-b-gray-400/40 px-6 py-4"
-      justify="between"
-    >
-      <Heading size={3}>{agent.name}</Heading>
-      {children}
-    </Flex>
-  );
-};
-
-const AGENT_MESSAGE_WIDTH_CLASS = 'max-w-[max(80%,350px)]';
-const USER_MESSAGE_WIDTH_CLASS = 'max-w-[80%]';
-
-type ChatBodyProps = {
-  agent: AiAgent;
-  messages: AiMessage[];
-  isThinking: boolean;
-  error: string | null;
-  sendMessage: (message: string) => void;
-  setIsAnimating: (animating: boolean) => void;
-};
-
-const ChatBody = ({ agent, messages, isThinking, error, sendMessage, setIsAnimating }: ChatBodyProps) => {
-  const previousMessageCountRef = useRef(messages.length);
-  
-  const handleAnimationComplete = useCallback(() => {
-    setIsAnimating(false);
-  }, [setIsAnimating]);
-  
-  useEffect(() => {
-    const lastMessage = messages[messages.length - 1];
-    const hasNewMessage = messages.length > previousMessageCountRef.current;
-    
-    if (hasNewMessage && lastMessage?.role === 'assistant') {
-      setIsAnimating(true);
-    }
-    
-    previousMessageCountRef.current = messages.length;
-  }, [messages, isThinking, setIsAnimating]);
-
-  const handleSuggestionClick = (suggestion: AiAgentStarterSuggestion) => {
-    sendMessage(suggestion.content);
-  };
-
-  const renderMessage = (message: AiMessage, index: number) => {
-    const isUser = message.role === 'user';
-    const isLastMessage = index === messages.length - 1;
-    const isLastAgentMessage = isLastMessage && !isUser;
-    
-    return (
-      <div 
-        key={message.id || index}
-        className={cn([
-          isUser
-            ? `ml-auto items-end bg-bg rounded-lg px-3 py-2 ${USER_MESSAGE_WIDTH_CLASS}`
-            : `mr-auto items-start ${AGENT_MESSAGE_WIDTH_CLASS}`
-        ])}
-      >
-        <MarkdownRenderer animate={isLastAgentMessage} onAnimationComplete={handleAnimationComplete}>
-          {message.content || ''}
-        </MarkdownRenderer>
-      </div>
-    );
-  };
-
-  return (
-    <Flex
-      direction="column"
-      className="grow overflow-auto bg-gray-50"
-    >
-      <ScrollArea.Root className="h-full grow-1 p-6">
-        <ScrollArea.Viewport className="h-full">
-          <Flex direction="column" gap={6}>
-            {agent.starterMessages?.map((msg, index) => (
-              <div
-                key={`message-${index}`}
-                className={cn("mr-auto items-start", AGENT_MESSAGE_WIDTH_CLASS)}
-              >
-                <MarkdownRenderer>
-                  {msg.content || ''}
-                </MarkdownRenderer>
-              </div>
-            ))}
-
-            {agent.starterSuggestions && messages.length === 0 && (
-              <Flex gap={2} wrap="wrap">
-                {agent.starterSuggestions.map((suggestion, index) => (
-                  <Button
-                    key={`suggestion-${index}`}
-                    variant="fill"
-                    mode="subtle"
-                    roundness="full"
-                    size="extrasmall"
-                    onPress={() => handleSuggestionClick(suggestion)}
-                  >
-                    {suggestion.label || suggestion.content}
-                  </Button>
-                ))}
-              </Flex>
-            )}
-
-            {messages.map((message, index) => renderMessage(message, index))}
-
-            {isThinking && (
-              <Text as="p" size={2} className="text-gray-500 mr-auto items-start">Thinking...</Text>
-            )}
-
-            {error && (
-              <Text as="p" size={2} className="bg-negative-100 text-negative-900 mx-auto items-start rounded-lg px-3 py-2">
-                Error: {error}
-              </Text>
-            )}
-          </Flex>
-        </ScrollArea.Viewport>
-        <ScrollArea.Scrollbar orientation="vertical" className="w-1 bg-gray-200 opacity-50">
-          <ScrollArea.Thumb className="flex-1 rounded-2xl bg-gray-400" />
-        </ScrollArea.Scrollbar>
-        <ScrollArea.Corner />
-      </ScrollArea.Root>
-    </Flex>
-  );
-};
-
-type ChatFooterProps = {
-  sendMessage: (message: string) => void;
-  isDisabled?: boolean;
-};
-
-const ChatFooter = ({ sendMessage, isDisabled = false }: ChatFooterProps) => {
-  const [inputValue, setInputValue] = useState('');
-
-  const handleSubmit = () => {
-    if (!inputValue.trim()) return;
-    sendMessage(inputValue);
-    setInputValue('');
-  };
-
-  const handleKeyDown = (e: KeyboardEvent<HTMLTextAreaElement>) => {
-    if (e.key === 'Enter' && !e.shiftKey) {
-      e.preventDefault();
-      handleSubmit();
-    }
-  };
-
-  return (
-    <div className="shrink-0 grow-0 border-t border-t-gray-400/40">
-      <Flex gap={2} align="end" className="w-full">
-        <div className="w-full [&_textarea]:!border-none [&_textarea]:!outline-none">
-          <TextArea
-            aria-label="message"
-            size="extralarge"
-            placeholder="Ask anything..."
-            value={inputValue}
-            onKeyDown={handleKeyDown}
-            onChange={(text) => setInputValue(text)}
-            roundness="none"
-            autogrow
-          />
-        </div>
-        <div className="mr-[7px] mb-[7px]">
-          <Button 
-            shape="square"
-            roundness="full"
-            variant="fill"
-            size="medium"
-            onPress={handleSubmit}
-            isDisabled={!inputValue.trim() || isDisabled}
-          >
-            <Send size={20} />
-          </Button>
-        </div>
-      </Flex>
-    </div>
-  )
-}
 
 export type { ChatProps };
