+++ packages/dashx-react/src/widgets/index.ts
@@ -1,4 +1,7 @@
 export { default as Chat } from './Chat.js';
+export { default as ChatBody } from './ChatBody.js';
+export { default as ChatFooter } from './ChatFooter.js';
+export { default as ChatHeader } from './ChatHeader.js';
 export { default as ChatPopover } from './ChatPopover.js';
 export { default as NotificationBell } from './NotificationBell.js';
 export { default as Preference } from './Preference.js';
